Fixed bugs and broken features
Debug class now working, now used in main and Robot. Fixed non working delay in Robot. Added raw reading to color sensor. Check now is working for every component in the robot, and show problem with the integrated LED of the board, and a Debug warning. Gained speed here and there. Now the robot should be way more reliable.